Logging in and displaying avatar location in sim
Client can now log in and display the avatar's current location within the sim.
Avatar is currently a capsule. Sim is assumed to have flat terrain at an elevation of 50 meters.
Also renamed OpenMetaverse.Vector3 to OpenMetaverse.OMVVector3 to make it less of a hassle to use Unity's own Vector3 in a script that is using OpenMetaverse in the header.