Beginning of avatar navigation
Camera is currently locked 5m behind the avatar. Avatar can walk forward and backward with W and S keys. Movement is sent as key presses instead of autonavigation. Movement is not smooth because interpolation and velocity are not yet implemented. Turning with A nd D keys is attempted but does not seem to be working correctly. Must investigate.