Major performance enhancement
Rezzed objects no longer have their own script that handles rezzing and rendering. Rather, that's all been moved into a RezzedObjectData class in SimManager. This results in massive gains in terms of performance.

An OctTree class has been added, but is not yet being used. Next step is to add object faces to it for implementing view distance properly.